backend/requirements.txt
fastapi
uvicorn
pydantic
opencv-python-headless
numpy
torch
torchvision
pillow
python-multipart
